Substituir atributo Long por int nas classes relacionadas

+++ src/main/java/br/com/zup/projeto_final/curtida/Curtida.java
@@ -13,8 +13,8 @@ public class Curtida {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_curtida;
-    private Long id_recurso;
+    private int id_curtida;
+    private int id_recurso;
     private String id_usuario;
     @Enumerated(EnumType.STRING)
     private Tipo tipo;
